refactor(Sales): moved sales endpoint mapping in Module
