+++ src/Sales/BrewUp.Sales.Facade/Endpoints/SalesEndpoints.cs
@@ -1,12 +1,31 @@
 ﻿using BrewUp.Sales.Facade.Validators;
 using BrewUp.Shared.Contracts;
 using FluentValidation;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Routing;
 
 namespace BrewUp.Sales.Facade.Endpoints;
 
 public static class SalesEndpoints
 {
+    public static IEndpointRouteBuilder MapSalesEndpoints(this IEndpointRouteBuilder endpoints) {
+        var group = endpoints.MapGroup("/v1/sales/")
+            .WithTags("Sales");
+
+        group.MapPost("/", SalesEndpoints.HandleCreateOrder)
+            .Produces(StatusCodes.Status400BadRequest)
+            .Produces(StatusCodes.Status201Created)
+            .WithName("CreateSalesOrder");
+
+        group.MapGet("/", SalesEndpoints.HandleGetOrders)
+            .Produces(StatusCodes.Status400BadRequest)
+            .Produces(StatusCodes.Status200OK)
+            .WithName("GetSalesOrders");
+
+        return endpoints;
+    }
+
     public static async Task<IResult> HandleCreateOrder(
         ISalesFacade salesFacade,
         IValidator<SalesOrderJson> validator,
